+++ packages/snap-preact-components/src/components/Molecules/CalloutBadge/CalloutBadge.stories.tsx
@@ -81,6 +81,16 @@ export default {
 			},
 			control: { type: 'boolean' },
 		},
+		limit: {
+			description: 'Number of badges per slot',
+			table: {
+				type: {
+					summary: 'number',
+				},
+				defaultValue: { summary: '1' },
+			},
+			control: { type: 'number' },
+		},
 		tag: {
 			description: 'Callout location tag',
 			table: {
@@ -95,10 +105,10 @@ export default {
 	},
 };
 
-const snapInstance = Snapify.search({ id: 'Result', globals: { siteId: '8uyt2m' } });
+const snapInstance = Snapify.search({ id: 'CalloutBadge', globals: { siteId: '8uyt2m' } });
 
 const ObservableCalloutBadge = observer(({ args, controller }: { args: CalloutBadgeProps; controller: SearchController }) => {
-	return <CalloutBadge {...args} result={controller?.store?.results[1] as Product} />;
+	return <CalloutBadge {...args} result={controller?.store?.results[0] as Product} />;
 });
 
 export const Default = (args: CalloutBadgeProps, { loaded: { controller } }: { loaded: { controller: SearchController } }) => {
@@ -118,14 +128,24 @@ Default.loaders = [
 
 			response.results[0].badges = [
 				{
-					tag: 'free-shipping-overlay',
+					tag: 'free-shipping',
 					value: 'Free Shipping',
 				},
-			];
-			response.results[1].badges = [
 				{
-					tag: 'free-shipping-callout',
-					value: 'Free Shipping',
+					tag: 'last-one',
+					value: 'Last One!',
+				},
+				{
+					tag: 'on-sale',
+					value: 'On Sale',
+				},
+				{
+					tag: 'save-percent',
+					value: 'Save 30%',
+				},
+				{
+					tag: 'inventory-remaining',
+					value: '1 in stock',
 				},
 			];
 
@@ -139,7 +159,12 @@ Default.loaders = [
 								name: 'Left',
 							},
 						],
-						right: [],
+						right: [
+							{
+								tag: 'right',
+								name: 'Right',
+							},
+						],
 						callout: [
 							{
 								tag: 'callout',
@@ -148,23 +173,53 @@ Default.loaders = [
 						],
 					},
 					tags: {
-						'free-shipping-overlay': {
-							location: 'left/left',
-							component: 'BadgeRectangle',
-							priority: 1,
-							enabled: true,
-							parameters: {
-								color: '#FF0000',
-								colorText: '#FFFFFF',
-							},
-						},
-						'free-shipping-callout': {
+						'free-shipping': {
 							location: 'callout/callout',
 							component: 'BadgeRectangle',
 							priority: 1,
 							enabled: true,
 							parameters: {
-								color: '#FF0000',
+								color: '#3A23AD',
+								colorText: '#FFFFFF',
+							},
+						},
+						'last-one': {
+							location: 'callout/callout',
+							component: 'BadgePill',
+							priority: 1,
+							enabled: true,
+							parameters: {
+								color: '#515151',
+								colorText: '#FFFFFF',
+							},
+						},
+						'inventory-remaining': {
+							location: 'callout/callout',
+							component: 'BadgePill',
+							priority: 1,
+							enabled: true,
+							parameters: {
+								color: '#382F5A',
+								colorText: '#FFFFFF',
+							},
+						},
+						'on-sale': {
+							location: 'left/left',
+							component: 'BadgePill',
+							priority: 1,
+							enabled: true,
+							parameters: {
+								color: '#00CEE1',
+								colorText: '#FFFFFF',
+							},
+						},
+						'save-percent': {
+							location: 'left/left',
+							component: 'BadgeRectangle',
+							priority: 1,
+							enabled: true,
+							parameters: {
+								color: '#8F6CF6',
 								colorText: '#FFFFFF',
 							},
 						},
@@ -172,6 +227,7 @@ Default.loaders = [
 				},
 			};
 		});
+
 		await snapInstance.search();
 
 		return {